pydantic move to core test
