+++ flytekit/__init__.py
@@ -216,7 +216,6 @@ if sys.version_info < (3, 10):
 else:
     from importlib.metadata import entry_points
 
-import flytekit.core.type_utils
 from flytekit._version import __version__
 from flytekit.configuration import Config
 from flytekit.core.array_node_map_task import map_task
