+++ flytekit/core/type_engine.py
@@ -1157,6 +1157,8 @@ class TypeEngine(typing.Generic[T]):
                 from flytekit.extras import pytorch  # noqa: F401
             if is_imported("sklearn"):
                 from flytekit.extras import sklearn  # noqa: F401
+            if is_imported("pydantic"):
+                from flytekit.extras import pydantic # noqa: F401
             if is_imported("pandas"):
                 try:
                     from flytekit.types.schema.types_pandas import PandasSchemaReader, PandasSchemaWriter  # noqa: F401
